+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+# Changes from version 0.8.2 to master 
+
+- Added `--docker.pull` option
+
 # Changes from version 0.8.1 to master 
 
 - Avoid using port offsets when using different `--starter.port`s that cause no overlap of port ranges.
@@ -6,6 +10,7 @@
 - After starters have bootstrapped, they elect a starter to be master over the cluster configuration.
   All changes (addition/removal) are forwarded to this master. 
   When the master is gone for too long, a new master is elected.
+
 # Changes from version 0.8.0 to master 
 
 - Fixed cluster setup in case where starters use different `--starter.port`s (#68).